Time markup
The first third of the book is dateable, but then an indeterminate amount of time follows so the remainder is time only.

There’s also an internal inconsistency in time: 19th June is referred to as a Tuesday in chapter IV, but a Friday in chapter II. Went with Tuesday as that fits the year before the book was published.